first success in rendering bucket models in Fabric
