+++ common/src/main/java/de/cech12/bucketlib/platform/Services.java
@@ -2,6 +2,7 @@ package de.cech12.bucketlib.platform;
 
 import de.cech12.bucketlib.CommonLoader;
 import de.cech12.bucketlib.platform.services.IBucketHelper;
+import de.cech12.bucketlib.platform.services.IClientHelper;
 import de.cech12.bucketlib.platform.services.IConfigHelper;
 import de.cech12.bucketlib.platform.services.IFluidHelper;
 import de.cech12.bucketlib.platform.services.IPlatformHelper;
@@ -18,6 +19,8 @@ public class Services {
 
     /** bucket helper instance */
     public static final IBucketHelper BUCKET = load(IBucketHelper.class);
+    /** Client helper instance */
+    public static final IClientHelper CLIENT = load(IClientHelper.class);
     /** Config helper instance */
     public static final IConfigHelper CONFIG = load(IConfigHelper.class);
     /** Fluid helper instance */
