+++ common/src/main/resources/assets/bucketlib/textures/item/lower_test_bucket.png
