+++ common/src/main/resources/assets/bucketlib/textures/item/test_bucket.png
