+++ common/src/main/resources/bucketlib.accesswidener
@@ -1,0 +1,3 @@
+accessWidener v2 named
+
+accessible method net/minecraft/client/renderer/item/BlockModelWrapper <init> (Lnet/minecraft/client/resources/model/BakedModel;Ljava/util/List;)V
