+++ fabric/src/main/java/de/cech12/bucketlib/api/crafting/BlockIngredient.java
@@ -24,6 +24,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Stream;
 
 public class BlockIngredient implements CustomIngredient {
 
@@ -73,7 +74,7 @@ public class BlockIngredient implements CustomIngredient {
     }
 
     @Override
-    public List<Holder<Item>> getMatchingItems() {
+    public Stream<Holder<Item>> getMatchingItems() {
         if (this.matchingStacks == null) {
             this.matchingStacks = new ArrayList<>();
             List<Block> blocks = new ArrayList<>();
@@ -97,7 +98,7 @@ public class BlockIngredient implements CustomIngredient {
                 });
             }
         }
-        return this.matchingStacks;
+        return this.matchingStacks.stream();
     }
 
     @Override
