+++ fabric/src/main/java/de/cech12/bucketlib/api/crafting/EntityIngredient.java
@@ -24,6 +24,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Stream;
 
 public class EntityIngredient implements CustomIngredient {
 
@@ -73,7 +74,7 @@ public class EntityIngredient implements CustomIngredient {
     }
 
     @Override
-    public List<Holder<Item>> getMatchingItems() {
+    public Stream<Holder<Item>> getMatchingItems() {
         if (this.matchingStacks == null) {
             this.matchingStacks = new ArrayList<>();
             List<EntityType<?>> entityTypes = new ArrayList<>();
@@ -104,7 +105,7 @@ public class EntityIngredient implements CustomIngredient {
                 });
             }
         }
-        return this.matchingStacks;
+        return this.matchingStacks.stream();
     }
 
     @Override
