--- fabric/src/main/java/de/cech12/bucketlib/client/model/UniversalBucketUnbakedModel.java
@@ -1,291 +1,0 @@
-package de.cech12.bucketlib.client.model;
-
-import com.google.common.collect.Maps;
-import de.cech12.bucketlib.api.BucketLib;
-import de.cech12.bucketlib.api.BucketLibTags;
-import de.cech12.bucketlib.api.item.UniversalBucketItem;
-import de.cech12.bucketlib.client.BucketLibClientMod;
-import de.cech12.bucketlib.mixin.BlockModelAccessor;
-import de.cech12.bucketlib.util.BucketLibUtil;
-import net.fabricmc.fabric.api.client.render.fluid.v1.FluidRenderHandlerRegistry;
-import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant;
-import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariantAttributes;
-import net.minecraft.client.multiplayer.ClientLevel;
-import net.minecraft.client.renderer.block.model.BakedOverrides;
-import net.minecraft.client.renderer.block.model.BakedQuad;
-import net.minecraft.client.renderer.block.model.BlockModel;
-import net.minecraft.client.renderer.block.model.ItemOverride;
-import net.minecraft.client.renderer.texture.MissingTextureAtlasSprite;
-import net.minecraft.client.renderer.texture.TextureAtlas;
-import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.client.resources.model.BakedModel;
-import net.minecraft.client.resources.model.BlockModelRotation;
-import net.minecraft.client.resources.model.Material;
-import net.minecraft.client.resources.model.ModelBaker;
-import net.minecraft.client.resources.model.ModelState;
-import net.minecraft.client.resources.model.UnbakedModel;
-import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.material.Fluid;
-import net.minecraft.world.level.material.Fluids;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.function.Function;
-
-/**
- * This implementation is based on net.neoforged.neoforge.client.model.DynamicFluidContainerModel.
- * Multiple changes were done to simplify the class
- */
-public class UniversalBucketUnbakedModel extends BlockModel implements UnbakedModel {
-
-    private static final Map<ResourceLocation, ResourceLocation> TEXTURE_MAP = Maps.newHashMap();
-    private static final Material MISSING_LOWER_CONTENT_MATERIAL = new Material(TextureAtlas.LOCATION_BLOCKS, getContentTexture(BucketLib.id("missing_lower_content")));
-
-    @NotNull
-    private Fluid fluid = Fluids.EMPTY;
-    @Nullable
-    private ResourceLocation otherContent = null;
-
-    private boolean isCracked = false;
-    private boolean isLower = false;
-
-    public UniversalBucketUnbakedModel(BlockModel blockModel) {
-        super(
-                ((BlockModelAccessor) blockModel).bucketlib_getParentLocation(),
-                blockModel.getElements(),
-                ((BlockModelAccessor) blockModel).bucketlib_getTextureMap(),
-                blockModel.hasAmbientOcclusion(),
-                blockModel.getGuiLight(),
-                blockModel.getTransforms(),
-                blockModel.getOverrides()
-        );
-    }
-
-    public UniversalBucketUnbakedModel(BlockModel blockModel, @NotNull Fluid fluid, @Nullable ResourceLocation otherContent, boolean isCracked, boolean isLower) {
-        this(blockModel);
-        this.fluid = fluid;
-        this.otherContent = otherContent;
-        this.isCracked = isCracked;
-        this.isLower = isLower;
-    }
-
-    /**
-     * Returns a new UniversalBucketModel with the given fluid.
-     */
-    public UniversalBucketUnbakedModel withFluid(Fluid newFluid, boolean isCracked) {
-        return new UniversalBucketUnbakedModel(this, newFluid, null, isCracked, false);
-    }
-
-    /**
-     * Returns a new UniversalBucketModel with the given other content.
-     */
-    public UniversalBucketUnbakedModel withOtherContent(ResourceLocation otherContent, boolean isCracked, boolean isLower) {
-        return new UniversalBucketUnbakedModel(this, Fluids.EMPTY, otherContent, isCracked, isLower);
-    }
-
-    private static ResourceLocation getContentTexture(ResourceLocation otherContentLocation) {
-        ResourceLocation texture = TEXTURE_MAP.get(otherContentLocation);
-        if (texture == null) {
-            String textureLocation = String.format("item/bucket_content/%s", otherContentLocation.getPath());
-            texture = otherContentLocation.withPath(textureLocation);
-            TEXTURE_MAP.put(otherContentLocation, texture);
-        }
-        return texture;
-    }
-
-    public ResourceLocation getParentLocation() {
-        return this.parentLocation;
-    }
-
-    @NotNull
-    @Override
-    public BlockModel getRootModel() {
-        if (this.parent == null) {
-            return this;
-        }
-        if (BucketLibClientMod.UNIVERSAL_BUCKET_MODEL.equals(this.getParentLocation())) {
-            //skip the BlockModel swap of the vanilla ItemRendering (ItemModelGenerator)
-            return this.parent;
-        }
-        return this.parent.getRootModel();
-    }
-
-    @NotNull
-    @Override
-    public BakedModel bake(@NotNull ModelBaker modelBaker, @NotNull Function<Material, TextureAtlasSprite> spriteGetter, @NotNull ModelState modelState) {
-        //resolve parents to use their defined materials, too
-        if (this.parent == null) {
-            resolveDependencies(BucketLibClientMod::getModel);
-        }
-        Material particleLocation = this.getMaterial("particle");
-        TextureAtlasSprite particleSprite = null;
-        if (isValid(particleLocation, spriteGetter)) {
-            particleSprite = spriteGetter.apply(particleLocation);
-        }
-
-        Material baseLocation = null;
-        if (this.isLower) {
-            if (this.isCracked) {
-                baseLocation = this.getMaterial("crackedLowerBase");
-            }
-            if (!isValid(baseLocation, spriteGetter)) {
-                baseLocation = this.getMaterial("lowerBase");
-            }
-        } else {
-            if (this.isCracked) {
-                baseLocation = this.getMaterial("crackedBase");
-            }
-            if (!isValid(baseLocation, spriteGetter)) {
-                baseLocation = this.getMaterial("base");
-            }
-        }
-
-        Material otherContentLocation = null;
-        Material fluidLocation = null;
-        Material fluidMaskLocation = null;
-        if (this.otherContent != null) {
-            otherContentLocation = new Material(TextureAtlas.LOCATION_BLOCKS, getContentTexture(this.otherContent));
-        }
-        if (this.fluid != Fluids.EMPTY) {
-            fluidLocation = new Material(TextureAtlas.LOCATION_BLOCKS, getContentTexture(BuiltInRegistries.FLUID.getKey(this.fluid)));
-            if (this.isCracked) {
-                fluidMaskLocation = this.getMaterial("crackedFluidMask");
-            }
-            if (!isValid(fluidMaskLocation, spriteGetter)) {
-                fluidMaskLocation = this.getMaterial("fluidMask");
-            }
-        }
-        //oversteer fluid texture if available
-        if (otherContent == null && isValid(fluidLocation, spriteGetter)) {
-            otherContentLocation = fluidLocation;
-        }
-
-        TextureAtlasSprite baseSprite = null;
-        if (isValid(baseLocation, spriteGetter)) {
-            baseSprite = spriteGetter.apply(baseLocation);
-            if (particleSprite == null)
-                particleSprite = baseSprite;
-        }
-        TextureAtlasSprite otherContentSprite;
-        if (isValid(otherContentLocation, spriteGetter)) {
-            otherContentSprite = spriteGetter.apply(otherContentLocation);
-            if (particleSprite == null)
-                particleSprite = otherContentSprite;
-        } else {
-            //if content texture is missing - fallback to pink content texture>
-            otherContentSprite = spriteGetter.apply(MISSING_LOWER_CONTENT_MATERIAL);
-        }
-
-        TextureAtlasSprite fluidSprite = null;
-        if (fluid != Fluids.EMPTY && FluidRenderHandlerRegistry.INSTANCE.get(fluid) != null) {
-            TextureAtlasSprite[] sprites = Objects.requireNotNull(FluidRenderHandlerRegistry.INSTANCE.get(fluid)).getFluidSprites(null, null, fluid.defaultFluidState());
-            if (sprites.length > 0) {
-                fluidSprite = sprites[0];
-                if (particleSprite == null)
-                    particleSprite = fluidSprite;
-            }
-        }
-
-        // if the fluid is lighter than air, will manipulate the initial state to be rotated 180deg to turn it upside down
-        if (fluid != Fluids.EMPTY && !fluid.defaultFluidState().is(BucketLibTags.Fluids.NO_FLIPPING) && FluidVariantAttributes.isLighterThanAir(FluidVariant.of(fluid))) {
-            modelState = BlockModelRotation.X180_Y0;
-        }
-
-        List<BakedQuad> quads = new ArrayList<>();
-
-        if (baseSprite != null) {
-            quads.addAll(GeometryUtils.bakeElements(this,
-                    GeometryUtils.createUnbakedItemElements(0, "base", baseSprite.contents()),
-                    baseSprite, modelState));
-        }
-        if (isValid(otherContentLocation, spriteGetter)) {
-            assert otherContentSprite != null;
-            quads.addAll(GeometryUtils.bakeElements(this,
-                    GeometryUtils.createUnbakedItemElements(-1, "content", otherContentSprite.contents()),
-                    otherContentSprite, modelState));
-        } else if (isValid(fluidMaskLocation, spriteGetter) && fluid != Fluids.EMPTY) {
-            TextureAtlasSprite templateSprite = spriteGetter.apply(fluidMaskLocation);
-            quads.addAll(GeometryUtils.bakeElements(this,
-                    GeometryUtils.createUnbakedItemMaskElements(1, "fluid", templateSprite.contents()),
-                    fluidSprite, modelState));
-        }
-
-        ContainedFluidOverrideHandler bakedOverrides = new ContainedFluidOverrideHandler(new BakedOverrides(modelBaker, List.of()), modelBaker, List.of(), this);
-        BakedModel bakedModel = new UniversalBucketBakedModel(quads, getTransforms(), bakedOverrides, particleSprite);
-        bakedOverrides.setBaseModel(bakedModel);
-
-        return bakedModel;
-    }
-
-    private boolean isValid(Material material, Function<Material, TextureAtlasSprite> spriteGetter) {
-        return material != null && !MissingTextureAtlasSprite.getLocation().equals(spriteGetter.apply(material).contents().name());
-    }
-
-    private static final class ContainedFluidOverrideHandler extends BakedOverrides {
-
-        private final Map<String, BakedModel> cache = Maps.newHashMap(); // contains all the baked models since they'll never change
-        private final BakedOverrides nested;
-        private BakedModel baseModel = null;
-        private final ModelBaker baker;
-        private final UniversalBucketUnbakedModel parent;
-
-        private Integer upperBreakTemperature = null;
-        private Integer lowerBreakTemperature = null;
-
-        private ContainedFluidOverrideHandler(BakedOverrides nested, ModelBaker baker, List<ItemOverride> itemOverrides, UniversalBucketUnbakedModel parent) {
-            super(baker, itemOverrides);
-            this.nested = nested;
-            this.baker = baker;
-            this.parent = parent;
-        }
-
-        private void setBaseModel(BakedModel baseModel) {
-            this.baseModel = baseModel;
-        }
-
-        @Nullable
-        @Override
-        public BakedModel findOverride(@NotNull ItemStack stack, @Nullable ClientLevel level, @Nullable LivingEntity entity, int seed) {
-            BakedModel overridden = nested.findOverride(stack, level, entity, seed);
-            if (overridden != null) return overridden;
-            if (stack.getItem() instanceof UniversalBucketItem bucket) {
-                boolean containsEntityType = false;
-                String content = BucketLibUtil.getEntityTypeString(stack);
-                if (content != null) {
-                    containsEntityType = true;
-                } else {
-                    content = BucketLibUtil.getContentString(stack);
-                }
-                Fluid fluid = null;
-                if (content == null) {
-                    fluid = BucketLibUtil.getFluid(stack);
-                    ResourceLocation location = BuiltInRegistries.FLUID.getKey(fluid);
-                    content = (location != BuiltInRegistries.FLUID.getDefaultKey()) ? location.toString() : null;
-                }
-                //reset cache if temperature config changed
-                if (!Objects.equals(upperBreakTemperature, bucket.getUpperBreakTemperature()) || !Objects.equals(lowerBreakTemperature, bucket.getLowerBreakTemperature())) {
-                    upperBreakTemperature = bucket.getUpperBreakTemperature();
-                    lowerBreakTemperature = bucket.getLowerBreakTemperature();
-                    cache.clear();
-                }
-                BakedModel bakedModel = cache.get(content);
-                if (bakedModel == null && content != null) {
-                    boolean isCracked = bucket.isCracked(stack);
-                    UniversalBucketUnbakedModel unbaked = (fluid == null) ? this.parent.withOtherContent(ResourceLocation.parse(content), isCracked, containsEntityType) : this.parent.withFluid(fluid, isCracked);
-                    bakedModel = unbaked.bake(baker, Material::sprite, BlockModelRotation.X0_Y0);
-                    cache.put(content, bakedModel);
-                }
-                return bakedModel;
-            }
-            return baseModel;
-        }
-    }
-
-}
