--- fabric/src/main/java/de/cech12/bucketlib/mixin/BlockModelAccessor.java
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package de.cech12.bucketlib.mixin;
-
-import net.minecraft.client.renderer.block.model.BlockModel;
-import net.minecraft.client.renderer.block.model.TextureSlots;
-import net.minecraft.client.resources.model.UnbakedModel;
-import net.minecraft.resources.ResourceLocation;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
-
-@Mixin(BlockModel.class)
-public interface BlockModelAccessor {
-	@Accessor("parent")
-	UnbakedModel bucketlib_getParent();
-
-	@Accessor("parentLocation")
-	ResourceLocation bucketlib_getParentLocation();
-
-	@Accessor("textureSlots")
-	TextureSlots.Data bucketlib_getTextureSlots();
-}
