+++ fabric/src/main/resources/bucketlib.fabric.mixins.json
@@ -6,9 +6,9 @@
   "mixins": [
   ],
   "client": [
-    "BlockModelAccessor",
     "FuelValuesMixin",
-    "ItemModelGeneratorAccessor"
+    "ItemModelGeneratorAccessor",
+    "ItemModelsMixin"
   ],
   "server": [
   ],
