+++ neoforge/src/main/java/de/cech12/bucketlib/platform/NeoforgeClientHelper.java
@@ -1,0 +1,26 @@
+package de.cech12.bucketlib.platform;
+
+import de.cech12.bucketlib.client.model.NeoforgeUniversalBucketModel;
+import de.cech12.bucketlib.client.model.UniversalBucketModel;
+import de.cech12.bucketlib.platform.services.IClientHelper;
+import net.minecraft.client.renderer.item.ItemModel;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.client.resources.model.SpriteGetter;
+import net.minecraft.world.level.material.Fluid;
+import net.neoforged.neoforge.client.ClientHooks;
+import net.neoforged.neoforge.client.extensions.common.IClientFluidTypeExtensions;
+import org.jetbrains.annotations.NotNull;
+
+public class NeoforgeClientHelper implements IClientHelper {
+
+    @Override
+    public TextureAtlasSprite getFluidTextureMaterial(SpriteGetter spriteGetter, Fluid fluid) {
+        return spriteGetter.get(ClientHooks.getBlockMaterial(IClientFluidTypeExtensions.of(fluid).getStillTexture()));
+    }
+
+    @Override
+    public UniversalBucketModel createItemModel(@NotNull UniversalBucketModel.Unbaked unbakedModel, @NotNull ItemModel.BakingContext bakingContext) {
+        return new NeoforgeUniversalBucketModel(unbakedModel, bakingContext);
+    }
+
+}
