+++ neoforge/src/main/resources/META-INF/services/de.cech12.bucketlib.platform.services.IClientHelper
@@ -1,0 +1,1 @@
+de.cech12.bucketlib.platform.NeoforgeClientHelper
